Split widget: timeline event row
