+++ extension/devtools/app/lib/main.dart
@@ -59,6 +59,7 @@ part 'widgets/batch_list.dart';
 part 'widgets/batch_header_row.dart';
 part 'widgets/batch_row.dart';
 part 'widgets/timeline_list.dart';
+part 'widgets/timeline_event_row.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
 part 'widgets/panel_state_cards.dart';
@@ -1104,53 +1105,6 @@ class _TimelinePanel extends StatelessWidget {
           ),
         );
       },
-    );
-  }
-}
-
-class _TimelineEventRow extends StatelessWidget {
-  const _TimelineEventRow({required this.event});
-
-  final TimelineEvent event;
-
-  @override
-  Widget build(BuildContext context) {
-    final tone = _timelineColors[event.type] ?? OrefPalette.teal;
-    return _GlassCard(
-      padding: const EdgeInsets.all(16),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Container(
-            width: 10,
-            height: 10,
-            margin: const EdgeInsets.only(top: 6),
-            decoration: BoxDecoration(color: tone, shape: BoxShape.circle),
-          ),
-          const SizedBox(width: 12),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  event.title,
-                  style: Theme.of(context).textTheme.bodyMedium,
-                ),
-                const SizedBox(height: 4),
-                Text(
-                  _formatTimelineDetail(event),
-                  style: Theme.of(context).textTheme.bodySmall,
-                ),
-              ],
-            ),
-          ),
-          const SizedBox(width: 12),
-          Text(
-            _formatAge(event.timestamp),
-            style: Theme.of(context).textTheme.bodySmall,
-          ),
-        ],
-      ),
     );
   }
 }
