+++ extension/devtools/app/lib/widgets/timeline_event_row.dart
@@ -1,0 +1,48 @@
+part of '../main.dart';
+
+class _TimelineEventRow extends StatelessWidget {
+  const _TimelineEventRow({required this.event});
+
+  final TimelineEvent event;
+
+  @override
+  Widget build(BuildContext context) {
+    final tone = _timelineColors[event.type] ?? OrefPalette.teal;
+    return _GlassCard(
+      padding: const EdgeInsets.all(16),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Container(
+            width: 10,
+            height: 10,
+            margin: const EdgeInsets.only(top: 6),
+            decoration: BoxDecoration(color: tone, shape: BoxShape.circle),
+          ),
+          const SizedBox(width: 12),
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  event.title,
+                  style: Theme.of(context).textTheme.bodyMedium,
+                ),
+                const SizedBox(height: 4),
+                Text(
+                  _formatTimelineDetail(event),
+                  style: Theme.of(context).textTheme.bodySmall,
+                ),
+              ],
+            ),
+          ),
+          const SizedBox(width: 12),
+          Text(
+            _formatAge(event.timestamp),
+            style: Theme.of(context).textTheme.bodySmall,
+          ),
+        ],
+      ),
+    );
+  }
+}
